Fixed compilation under GradleRIO and made it more idiomatic
There's some FIXME comments on stuff that we need to fix in WPILib's API.

Empty overidden functions could also be removed completely, but I left
them there because I figured you did that for a reason.

If we made the static variables members instead, we could eliminate the
duplication of putting definitions in the .cpp, but we'd need a way to get
the Robot instance in that case.

The wpiformat files could be removed if we want. I put them in there to
make cleaning this up easier for me.